Split RealEstate and Building models into their own files
