+++ backend/hitas/models/__init__.py
@@ -1,3 +1,5 @@
+from hitas.models.building import Building
 from hitas.models.codes import BuildingType, Developer, FinancingMethod, PostalCode
-from hitas.models.housing_company import Building, HousingCompany, RealEstate
+from hitas.models.housing_company import HousingCompany, HousingCompanyState
 from hitas.models.property_manager import PropertyManager
+from hitas.models.real_estate import RealEstate
